refactor: extract DMas interface
